requirements.txt
bm25s
PyStemmer
fastapi
uvicorn[standard]
